전남대 BE_박진화 4주차 과제 (2단계) (#312)
* feat: migrate prev project codes

* docs: delete unnecessary lines

* docs: add README

* docs: modify typo

* rename: move prev codes to oldTestCodes directory

* feat: add codes about Category

* feat: changed feature about set to update

* refactor: modify ProductController as modified Product entity

* feat: add getEmailFromRequest method to JwtUtil

* refactor: update controllers with clear return types and improve readability

* feat: add new fields(color, imageUrl, description)  and update Category entity

* feat: add path variable support for category deletion

* feat: integrate category handling in ProductController

* feat: update Product entity to include category handling

* feat: update ProductRequest, ProductResponse, ProductService to include category handling

* feat: Add code to WishController to verify ProductId

* "remove: delete unnecessary product.html"

* docs: update README

* feat: make addOption page for admin

* refactor: change dependency injection from field to constructor

* feat: implement option features

* feat: add option handle feature

* refactor: change dependency injection from field to constructor (JwtAuthenticationFilter, WebConfig, WishService)

* feat: add option schema

* feat: add redundancy option check feature

* feat: add properties

* refactor: change annotation from korean to english

* Revert "feat: add redundancy option check feature"

This reverts commit b74f7ad6399e7319146f39a3a94be2e2992cea06.

* Revert "Revert "feat: add redundancy option check feature""

This reverts commit cd2982f3e76927c6427344062c53dfb5542342b8.

* feat: add enforcing product have at least one option

* feat: add feat ProductController, OptionDTO, ProductRequest

ProductController: when adding Product, check Option list is empty or
not, using ProductValidator
OptionDTO: add annotation checking validation
ProductRequest: make method possible to include multiple options and to
access option list as adding setOption methods.

* fix: make GlobalExceptionHandler handles proper error

* refactor: changes id variable to indicate clearly

* feat: Apply review feedback and add test code

* feat: add builder, delete unnecessary import and annotations

* fix: modify as it works on test codes, change order, add option on data

* feat: add CustomException, delete unnecessary annotations

* rename: modify OptionDTO to OptionRequest

* refactor: Improve validation, error handling, entity retrieval

- Removed unnecessary import or annotation
- added validation error handling with bindingResult
- changed error handling to use CustomException
- Replaced Optional with entity retrieval

* refactor: change to use CustomException

* refactor: change to use CustomException

* refactor: using SpringBootTest to WebMvcTest

* refactor: apply renamed OptionRequest

* refactor: enhance validation

+++ src/main/java/gift/service/ProductService.java
@@ -2,9 +2,9 @@ package gift.service;
 
 import gift.entity.Category;
 import gift.entity.Product;
+import gift.exception.CustomException;
 import gift.repository.CategoryRepository;
 import gift.repository.ProductRepository;
-import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -26,7 +26,7 @@ public class ProductService {
     }
 
     public Product findById(Long ProductId) {
-        return productRepository.findById(ProductId).orElseThrow(()-> new EntityNotFoundException("Product not found"));
+        return productRepository.findById(ProductId).orElseThrow(()-> new CustomException.EntityNotFoundException("Product not found"));
     }
 
     public Product save(Product product) {
@@ -34,12 +34,15 @@ public class ProductService {
     }
 
     public void deleteById(Long ProductId) {
+        if(!productRepository.existsById(ProductId)) {
+            throw new CustomException.EntityNotFoundException("Product not found");
+        }
         productRepository.deleteById(ProductId);
     }
 
     public Category getCategoryById(Long categoryId) {
         return categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new EntityNotFoundException("Category not found"));
+                .orElseThrow(() -> new CustomException.EntityNotFoundException("Category not found"));
     }
 }
